Changed stop to return if subject isn't defined. Reintroduced cron job logging statement as that's not the cause of the observed issue.

+++ packages/deep-microservice-collection/src/command/k8s-job.mjs
@@ -59,6 +59,9 @@ class K8sJob extends Command {
     }
 
     async stop() {
+
+        if (!this._obj) return;
+
         try {
             this._logger.info(`Deleting job.`);
             await this._k8sClient.delete(this._obj);
